Move tests to new layout
